+++ packages/kit/test/apps/basics/test/client.test.js
@@ -827,32 +827,56 @@ test.describe('Invalidation', () => {
 test.describe('data-sveltekit attributes', () => {
 	test('data-sveltekit-preload-code', async ({ page }) => {
 		/** @type {string[]} */
-		const requests = [];
-		page.on('request', (r) => {
-			requests.push(r.url());
+		const responses = [];
+
+		const nodes_location = process.env.DEV
+			? '.svelte-kit/generated/client/nodes/'
+			: '/_app/immutable/nodes/';
+
+		page.on('response', async (response) => {
+			const url = response.url();
+			if (url.includes(nodes_location)) {
+				responses.push(url);
+			}
 		});
 
 		// eager
 		await page.goto('/data-sveltekit/preload-code');
-		expect(requests.length).toBeGreaterThanOrEqual(1);
+		await page.locator('#eager').hover();
+		await page.locator('#eager').dispatchEvent('touchstart');
+		// expect 4 nodes on initial load: root layout, root error, current page, and eager preload
+		expect(responses.length).toEqual(4);
 
 		// viewport
-		requests.length = 0;
+		responses.length = 0;
 		page.locator('#viewport').scrollIntoViewIfNeeded();
-		await Promise.all([page.waitForTimeout(100), page.waitForLoadState('networkidle')]);
-		expect(requests.length).toBeGreaterThanOrEqual(1);
+		await page.locator('#viewport').hover();
+		await page.locator('#viewport').dispatchEvent('touchstart');
+		await Promise.all([
+			page.waitForTimeout(100), // wait for preloading to start
+			page.waitForLoadState('networkidle') // wait for preloading to finish
+		]);
+		expect(responses.length).toEqual(1);
 
 		// hover
-		requests.length = 0;
-		await page.locator('#hover').dispatchEvent('mousemove');
-		await Promise.all([page.waitForTimeout(100), page.waitForLoadState('networkidle')]);
-		expect(requests.length).toBeGreaterThanOrEqual(1);
+		responses.length = 0;
+		await page.locator('#hover').hover();
+		await page.locator('#hover').dispatchEvent('touchstart');
+		await Promise.all([
+			page.waitForTimeout(100), // wait for preloading to start
+			page.waitForLoadState('networkidle') // wait for preloading to finish
+		]);
+		expect(responses.length).toEqual(1);
 
 		// tap
-		requests.length = 0;
+		responses.length = 0;
+		await page.locator('#tap').hover();
 		await page.locator('#tap').dispatchEvent('touchstart');
-		await Promise.all([page.waitForTimeout(100), page.waitForLoadState('networkidle')]);
-		expect(requests.length).toBeGreaterThanOrEqual(1);
+		await Promise.all([
+			page.waitForTimeout(100), // wait for preloading to start
+			page.waitForLoadState('networkidle') // wait for preloading to finish
+		]);
+		expect(responses.length).toEqual(1);
 	});
 
 	test('data-sveltekit-preload-data', async ({ page }) => {
@@ -863,42 +887,59 @@ test.describe('data-sveltekit attributes', () => {
 				req
 					.response()
 					.then(
-						(res) => res.text(),
+						(res) => res?.text(),
 						() => ''
 					)
-					.then((response) => {
-						if (response.includes('this string should only appear in this preloaded file')) {
+					.then((text) => {
+						if (text?.includes('this string should only appear in this preloaded file')) {
 							requests.push(req.url());
 						}
 					});
 			}
+
+			if (req.url().includes('__data.json')) {
+				requests.push(req.url());
+			}
 		});
 
 		await page.goto('/data-sveltekit/preload-data');
-		await page.locator('#one').dispatchEvent('mousemove');
+		await page.locator('#one').hover();
+		await page.locator('#one').dispatchEvent('touchstart');
 		await Promise.all([
 			page.waitForTimeout(100), // wait for preloading to start
 			page.waitForLoadState('networkidle') // wait for preloading to finish
 		]);
-		expect(requests.length).toBe(1);
+		expect(requests.length).toBe(2);
 
 		requests.length = 0;
 		await page.goto('/data-sveltekit/preload-data');
-		await page.locator('#two').dispatchEvent('mousemove');
+		await page.locator('#two').hover();
+		await page.locator('#two').dispatchEvent('touchstart');
 		await Promise.all([
 			page.waitForTimeout(100), // wait for preloading to start
 			page.waitForLoadState('networkidle') // wait for preloading to finish
 		]);
-		expect(requests.length).toBe(1);
+		expect(requests.length).toBe(2);
 
 		requests.length = 0;
 		await page.goto('/data-sveltekit/preload-data');
-		await page.locator('#three').dispatchEvent('mousemove');
+		await page.locator('#three').hover();
+		await page.locator('#three').dispatchEvent('touchstart');
 		await Promise.all([
 			page.waitForTimeout(100), // wait for preloading to start
 			page.waitForLoadState('networkidle') // wait for preloading to finish
 		]);
 		expect(requests.length).toBe(0);
+
+		requests.length = 0;
+		await page.goto('/data-sveltekit/preload-data');
+		await page.locator('#tap').hover();
+		await page.locator('#tap').dispatchEvent('touchstart');
+		await Promise.all([
+			page.waitForTimeout(100), // wait for preloading to start
+			page.waitForLoadState('networkidle') // wait for preloading to finish
+		]);
+		expect(requests.length).toBe(2);
 	});
 
 	test('data-sveltekit-preload-data network failure does not trigger navigation', async ({
@@ -965,6 +1006,47 @@ test.describe('data-sveltekit attributes', () => {
 		]);
 
 		expect(page).toHaveURL('/data-sveltekit/preload-data/offline/slow-navigation');
+	});
+
+	test('data-sveltekit-preload-data tap works after data-sveltekit-preload-code hover', async ({
+		page
+	}) => {
+		/** @type {string[]} */
+		const requests = [];
+		page.on('request', (req) => {
+			if (req.resourceType() === 'script') {
+				req
+					.response()
+					.then(
+						(res) => res?.text(),
+						() => ''
+					)
+					.then((text) => {
+						if (text?.includes('this string should only appear in this preloaded file')) {
+							requests.push(req.url());
+						}
+					});
+			}
+
+			if (req.url().includes('__data.json')) {
+				requests.push(req.url());
+			}
+		});
+
+		await page.goto('/data-sveltekit/preload-data');
+		await page.locator('#hover-then-tap').hover();
+		await Promise.all([
+			page.waitForTimeout(100), // wait for preloading to start
+			page.waitForLoadState('networkidle') // wait for preloading to finish
+		]);
+		expect(requests.length).toBe(1);
+
+		await page.locator('#hover-then-tap').dispatchEvent('touchstart');
+		await Promise.all([
+			page.waitForTimeout(100), // wait for preloading to start
+			page.waitForLoadState('networkidle') // wait for preloading to finish
+		]);
+		expect(requests.length).toBe(2);
 	});
 
 	test('data-sveltekit-reload', async ({ baseURL, page, clicknav }) => {
@@ -1365,6 +1447,14 @@ test.describe('reroute', () => {
 		await page.click("a[href='/reroute/basic/a']");
 		expect(await page.textContent('h1')).toContain(
 			'Successfully rewritten, URL should still show a: /reroute/basic/a'
+		);
+	});
+
+	test('Apply async reroute during client side navigation', async ({ page }) => {
+		await page.goto('/reroute/async');
+		await page.click("a[href='/reroute/async/a']");
+		expect(await page.textContent('h1')).toContain(
+			'Successfully rewritten, URL should still show a: /reroute/async/a'
 		);
 	});